backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
sqlalchemy[mypy]
psycopg2-binary
